+++ waspc/data/Generator/templates/server/src/routes/auth/passport/generic/configMapping.js
